refactor(mastering): extract Stages 3+4 (pre-QC + mastering) (#290 D5)
